# openapi.yaml
openapi: 3.0.3
info:
  title: Moms Website API
  version: 1.0.0
  description: >
    OpenAPI specification for the current Moms Website backend.
    Base route prefix is /api.
servers:
  - url: http://localhost:5001/api
    description: Local development
...
paths:
...
  /appointments:
    get:
      tags: [Appointments]
      summary: List appointments
      operationId: listAppointments
      parameters:
        - in: query
          name: dateFrom
          schema:
            type: string
            format: date-time
        - in: query
          name: dateTo
          schema:
            type: string
            format: date-time
        - in: query
          name: serviceId
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Appointment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        "500":
          $ref: "#/components/responses/InternalServerError"
    post:
      tags: [Appointments]
      summary: Create appointment
      description: >
        Enforces scheduling policy (Mon-Fri, 9am-5pm MST, 24h advance),
        overlap protection, and reminder scheduling.
      operationId: createAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentCreateRequest"
      responses:
        "201":
          description: Appointment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Service not found
        "409":
          description: Scheduling conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: Selected time conflicts with an existing appointment
        "500":
          $ref: "#/components/responses/InternalServerError"
  /appointments/available:
    get:
      tags: [Appointments]
      summary: Get available slots by service and date
      operationId: getAvailableAppointmentSlots
      parameters:
        - in: query
          name: serviceId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: date
          required: true
          schema:
            type: string
            pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$"
            example: 2026-05-15
      responses:
        "200":
          description: Available slot list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlotsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Service not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /appointments/{id}:
    get:
      tags: [Appointments]
      summary: Get appointment by ID
      operationId: getAppointmentById
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Appointment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      tags: [Appointments]
      summary: Update appointment
      description: >
        Supports customer data and date changes; date changes enforce 24h policy and overlap checks.
      operationId: updateAppointment
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdateRequest"
      responses:
        "200":
          description: Appointment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Scheduling conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      tags: [Appointments]
      summary: Cancel appointment
      description: Requires cancellation at least 24 hours in advance.
      operationId: deleteAppointment
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Appointment deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /appointments/{id}/confirm:
    put:
      tags: [Appointments]
      summary: Confirm appointment after payment
      description: Requires server-trusted source outside test mode.
      operationId: confirmAppointment
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentConfirmRequest"
      responses:
        "200":
          description: Appointment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Unauthorized payment confirmation source
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-admin-key
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: Resource identifier
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            oneOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - $ref: "#/components/schemas/MessageResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        bufferMinutes:
          type: integer
        isPublished:
          type: boolean
        clientId:
          type: string
          format: uuid
          nullable: true
      required:
        [id, title, description, price, durationMinutes, bufferMinutes, isPublished]
...
    Appointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientFirstName:
          type: string
        clientLastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
        date:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
          nullable: true
        timezone:
          type: string
          example: MST
        serviceId:
          type: string
          format: uuid
        states:
          type: string
          enum: [pending, confirmed, cancelled]
        paymentStatus:
          type: string
        paymentMethod:
          type: string
          nullable: true
        tipAmount:
          type: number
          nullable: true
        paymentProvider:
          type: string
          nullable: true
        paymentExternalId:
          type: string
          nullable: true
        calendarEventId:
          type: string
          nullable: true
        cancelledAt:
          type: string
          format: date-time
          nullable: true
        rescheduledAt:
          type: string
          format: date-time
          nullable: true
        service:
          $ref: "#/components/schemas/Service"
      required:
        - id
        - clientFirstName
        - clientLastName
        - email
        - date
        - timezone
        - serviceId
        - states
    AppointmentCreateRequest:
      type: object
      required:
        [clientFirstName, clientLastName, email, phone, date, serviceId]
      properties:
        clientFirstName:
          type: string
        clientLastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        date:
          type: string
          format: date-time
        serviceId:
          type: string
          format: uuid
        paymentMethod:
          type: string
          enum: [venmo, credit_card]
        paymentStatus:
          type: string
          enum: [pending, paid]
        tipAmount:
          type: number
    AppointmentUpdateRequest:
      type: object
      minProperties: 1
      properties:
        clientFirstName:
          type: string
        clientLastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        date:
          type: string
          format: date-time
        states:
          type: string
          enum: [pending, confirmed, cancelled]
        paymentMethod:
          type: string
          enum: [venmo, credit_card]
        paymentStatus:
          type: string
          enum: [pending, paid]
        tipAmount:
          type: number
    AppointmentConfirmRequest:
      type: object
      required: [appointmentId, paymentStatus]
      properties:
        appointmentId:
          type: string
          format: uuid
        paymentStatus:
          type: string
          enum: [completed, pending, failed]
...
    AvailabilitySlotsResponse:
      type: object
      properties:
        date:
          type: string
          example: 2026-05-15
        timezone:
          type: string
          example: MST
        serviceId:
          type: string
          format: uuid
        durationMinutes:
          type: integer
        bufferMinutes:
          type: integer
        slots:
          type: array
          items:
            type: string
            format: date-time
      required: [date, timezone, serviceId, durationMinutes, bufferMinutes, slots]
...